feeds-master/requirements.txt
cffi
ffmpeg-python
google-cloud-speech
pydub
requests
twilio
webrtcvad
python-dotenv
flask
pyaudio